feat: make Cmd+Shift+L (cycle theme) shortcut app-wide
Move the theme-cycle shortcut and cycleTheme() out of MailboxView into the
useTheme composable, and register the Cmd/Ctrl+Shift+L keydown globally in
App.vue so it works on every page, not just the mailbox.
